move targetIsStage into a standalone component for costume tab text
targetIsStage triggers a render of most of GUI currently. But it is
only needed to be able to determine the costume tab text. As a
standalone component we can avoid a lot of time render much of the GUI
that does not need to be. The other parts of the gui that do need to
update because the editing target change will continue to do so.